@@ -2,6 +2,7 @@
 using Microsoft.Extensions.Logging;
 using StockParser.Common;
 using StockParser.Domain;
+using StockParser.Domain.Models;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
@@ -11,25 +12,27 @@ namespace StockParser.Data.WebParser
 {
     public class BigParaParser:IWebParser
     {
-        public BigParaParser(ICustomLogger logger)
+        public BigParaParser(
+            //ICustomLogger logger
+            )
         {
-            _logger = logger;
+            //_logger = logger;
         }
-        private HashSet<StockDto> StockData;
+        private HashSet<IBistStock> StockData;
 
-        public ICustomLogger _logger { get; }
+        //public ICustomLogger _logger { get; }
 
-        public async Task<ServiceResult<HashSet<StockDto>>> GetStockData()
+        public async Task<ServiceResult<HashSet<IBistStock>>> GetStockData()
         {
             try
             {
-                _logger.LogInformation("GetStockData function called");
+                //_logger.LogInformation("GetStockData function called");
                 if (StockData == null)
                 {
-                    StockData = new HashSet<StockDto>();
+                    StockData = new HashSet<IBistStock>();
                     var mainUrl = "http://push.bigpara.com/borsa/hisse-fiyatlari/";
                     var letterlist = new List<string> { "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "R", "S", "T", "U", "V", "Y", "Z" };
-                    var taskList = new List<Task<HashSet<StockDto>>>();
+                    var taskList = new List<Task<HashSet<IBistStock>>>();
                     foreach (var letter in letterlist)
                     {
                         var url = new Uri(new Uri(mainUrl), $"{letter}-harfi-ile-baslayan-hisseler");
@@ -38,7 +41,7 @@ namespace StockParser.Data.WebParser
 
                     await Task.WhenAll(taskList);
 
-                    _logger.LogInformation("All pages parsed");
+                    //_logger.LogInformation("All pages parsed");
                     foreach (var task in taskList)
                     {
                         var stocks = await task;
@@ -50,21 +53,21 @@ namespace StockParser.Data.WebParser
                     }
                     //_logger.LogTrace($"Total {StockData.Count} stocks parsed");
                 }
-                return new ServiceResult<HashSet<StockDto>>(ServiceStatus.Ok, StockData);
+                return new ServiceResult<HashSet<IBistStock>>(ServiceStatus.Ok, StockData);
             }
             catch (Exception ex)
             {
-                _logger.LogError("GetStockData:"+ex.Message);
-                return new ServiceResult<HashSet<StockDto>>(ServiceStatus.Error, null,"GetStockData hata: "+ex.Message);
+                //_logger.LogError("GetStockData:"+ex.Message);
+                return new ServiceResult<HashSet<IBistStock>>(ServiceStatus.Error, null,"GetStockData hata: "+ex.Message);
             }
 
         }
 
-        public async Task<HashSet<StockDto>> GetDataPerPage(Uri url)
+        public async Task<HashSet<IBistStock>> GetDataPerPage(Uri url)
         {
             try
             {
-                _logger.LogInformation("GetDataPerPage function called : " + url.ToString());
+                //_logger.LogInformation("GetDataPerPage function called : " + url.ToString());
                 HttpClient client = new HttpClient();
                 var response = await client.GetAsync(url);
                 var pageContents = await response.Content.ReadAsStringAsync();
@@ -74,10 +77,10 @@ namespace StockParser.Data.WebParser
 
                 var tableDiv = pageDocument.DocumentNode.SelectNodes("//div[@class='tableCnt']//div[@class='tBody']//ul");
 
-                var stockList = new List<StockDto>();
+                var stockList = new List<IBistStock>();
                 foreach (var row in tableDiv)
                 {
-                    var stock = new StockDto();
+                    var stock = new IBistStock();
                     var list = new List<string>();
                     var coloums = row.SelectNodes("li");
                     foreach (var coloum in coloums)
@@ -104,6 +107,7 @@ namespace StockParser.Data.WebParser
                     stock.HighestPrice = numbers[3];
                     stock.LowestPrice = numbers[4];
                     stock.AveragePrice = numbers[5];
+                    stock.Date = DateTime.UtcNow.Date;
 
                     if (long.TryParse(list[7].Replace(".", ""), out long lot))
                     {
@@ -115,11 +119,11 @@ namespace StockParser.Data.WebParser
                     }
                     stockList.Add(stock);
                 }
-                return new HashSet<StockDto>(stockList);
+                return new HashSet<IBistStock>(stockList);
             }
             catch (Exception ex)
             {
-                _logger.LogError("Exception:"+ex.Message);
+                //_logger.LogError("Exception:"+ex.Message);
                 return null;
             }
            
