@@ -1,5 +1,6 @@
 ﻿using StockParser.Common;
 using StockParser.Domain;
+using StockParser.Domain.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -7,6 +8,6 @@ namespace StockParser.Data.WebParser
 {
     public interface IWebParser
     {
-        Task<ServiceResult<HashSet<StockDto>>> GetStockData();
+        Task<ServiceResult<HashSet<IBistStock>>> GetStockData();
     }
 }
